Config system
The description is in the README. The only thing not mentioned is why
I'm removing the htmlmin task. With this config stuff, we're using ESI,
which relies on comments and custom tags not getting mangled by htmlmin.
Perhaps if I later learn about desirable benefits that running htmlmin
brings, I'll find a way to make them work together.